Added configuration in configuration for enabling and disabling public_ntp serving. This differs from ntp_server which is required for OCN <> CN communication. If packets are received with a private key then ntp requests are assumed to come from CN
The public_ntp flag can be changed while radclock is running through process signaling which is executed via ansible.
This signal activates a light reread of the conf file which only checks the public_ntp flag.
This avoids the need to close running threads and restart them
It also makes sure that the last public_ntp setting called from ansible is maintained even between radclock restarts as it pulls from the radclock.conf