Thread fix (#6)
Reengineering of TRIGGER and DATA_PROC thread structure and packet sending/sinking.

The tag-team  TRIGGER <--> DATA_PROC  thread structure was tuned for performance
in the single-server case. With multiple servers, grid points could be missed by
TRIGGER due to excessive waiting and blocking, and this could become systematic,
starving some clocks for extended periods or even permanently.

Several significant changes were made to eliminate this problem:
 - creation of a FIFO queue for alarms to ensure none ever missed
 - decoupling of the threads to run independently to kill one source of blocking
 - moving from blocking to non-blocking sockets when grid points are too dense
Details follow.
This commit is joint work with Rohan Turai.

Thread management:
Removal of the TRIGGER <--> DATA_PROC signaling, and associated mutexes
and soft-signaling variables, a big simplification!  The threads now
run independently without any signaling, neither explicit nor implied.
When timeouts are used in TRIGGER, this is now purely to support retry code,
not to try to time the awakening of PROC.

MAIN:
The handler catch_alarm now places Alarm signals into a FIFO queue when they are
received, before signaling TRIGGER. FIFO support is defined in new files FIFO.{h,c}.
A drop HEAD dropping policy is used which ensures that in the extreme case of
overflow, the more important more recent packets are given priority.


TRIGGER:
Now gives up its timeslice only when waiting for the alarm of the next grid point.
Previously it also gave it up when waiting for a signal from PROC to proceed.

The FIFO queue is initialized with FIFO_init in ntp_client_init, and destroyed
with FIFO_destroy as part of thread cleanup in thread_trigger.

When the risk of missing the arrival of the next grid point while processing the
current one is non-negligible, we use non-blocking recvfrom to quickly send the
packet then return to the alarm wait position. Here a request pkt from server s is
`sunk` the next time a s-grid point is processed, but this delay does not impact
stamp processing in PROC.
When the risk is very low, retries are allowed, using blocking recvfrom as
before to catch the matching returning response. Non-blocking recvfrom is used
to clear out (sink) spurious retries if they occur.
Details are given in comments in client_ntp.c .

Currently the heuristic defining "non-negligible" is
if (adjusted_period < 4 || gridgap < 3)
which seems to work well both for NTC conditions (many servers, poll_period = 1),
and traditional (single server, large poll_period), and more generally.
The logic here is assuming a common poll_period for all servers still.

The server grids are now offset evenly over the entire polling period. 

DATA_PROC:
Now gives up its timeslice based on usleep(pktwait) in thread_data_processing,
called once there is no further raw data to process.
Previously it gave it up when passing control to TRIGGER to send the next packet.

The wait time is based on a simple heuristic: being just the estimated
average inter-grid interval (all servers superimposed), clipped at 1 second.
This caters for both NTC conditions where new packets to process are expected on
the grid timescale, and allows for regular waking up for starvation checking and
telemetry updating when grid points are widely spaced.

For efficiency, the retry loop within get_network_stamp now exits as soon as the
raw data buffer is emptied. It is no longer important to try to wait until the
response associated to the most recent grid point arrives: this is too hard to predict,
and in any event in the worst case a stamp`s delivery to the algo will now not be
delayed by more than a second.

Misc:
- Stamp matching verbosity improvement, and alignment of some stamp matching code
  to the better version under OCNdev.  Relevant to do this here as it was part of
  the testing of the changes.
- Updates to RADclockDoc_threads.txt to reflect this commit, and limited other
  updates.
- Update to AUTHORS file:  including Jason Traish and Rohan Saha Turai