Inclusion of Workflow scripts
A new Workflow_scripts directory containing scripts supporting my workflow on:
  VM:  for amd64, FreeBSD or Linux
  RPi: for ARM-Pi, headless operation .
The workflow includes both FF kernel and userland radclock aspects.

These are for use by other project members to seed/adopt-into their own
workflows as they see fit. The scripts include the use of earthly,
though that is not fully mature in the case of VM.

The readme.md file provides a fair amount of detail on what is provided.