Port of FFcode from FreeBSD 12.1 --> 12.3
The port was straightforward.
The largest OS change was in kern_tc.c, where new macros and functions
appeared enabling more careful access and setting of bintime fields in timehands,
used in time getting and reading functions like fbclock_bin{,up}time  get*time .
Chose not to port this to the analogous FF functions in kern_ffclock.c at this
point - not clear is needed.

- minor white space style consistency tweaks
- minor changes/fixed in test programs test_{FFclocks,clockcompare}.c