NTC: Establishment of Asymmetry Calibration system
Code to enable asymmetry calibration, based on system clock
timestamps taken on the sync-flow packets, assumed to be synchronized,
available thanks to the STAMP_NTP_INT matching mode (create_stamp.c).
The main use case is for an OCN running live, using a FB system clock
synchronized as a stratum-1 via a PPS input available during a limited
calibration period.

The system incorporates a new approach to online Up and Down shift
detection applied to both Db and Df, and a robust calibration
algorithm sitting on top, separately for each server.
At the end of the calibration period, the best estimate (and the
time it pertains to) are made available. They can be used in future
to improve accuracy and path tracking of the corresponding RADclocks,
in particular the asymmetry management and asymmetry-tracking sub-algos.
Calibration can be initiated and terminated via the light reload
mechanism (signalled by SIGUSR2).

Though uncoupled to the sync algo right now, most of the code is within
sync_bidir.{c,h} as is called within process_stamp. This
is also appropriate as it will very likely form the basis of the next
generation of the asymmetry tracking sub-algo.

Calibration control :
Enabled by a new boolean conf file member “asym_calibrate”
that should be deactivated by default (=off).
When activated via light reloading, calibration state is initialized via
radclock_main.c:create_calib (within signal_handler), and
handle->calibrate is set to 1.
When deactivated, handle->calibrate = 2 to signal shutdown, this is then
acted on in process_stamp for all servers (independently of stamp arrival,
and so cannot be blocked even if all servers starved) and
it is reset to 0 when this is complete.
During live calibration, the IntRef members of stamp_t will
be populated (update_stamp_queue), otherwise they will be zero
(when running dead, they will reflect the input data).
For dead input calibration cannot be activated, but hack activation code
in process_stamp can be uncommented to control activation/deactivation.
If ascii output is selected, this enables the calibration period
to be clearly seen via the IntRef columns ‘switching on and off’.

New state variables  (sync_algo.h)
  struct bidir_calibstate :
    - state for the calibration algo itself
    - owner of struct BLtrack_state (see next), for each of Df, Db
  struct BLtrack_state  :
    - state for the BL tracking algo BL_track() for scalar input
    - also holds needed histories:  for input data, and inferred BL
      so that it can be backdated (currently not used by calibration,
      and ascii output is pushed per-stamp and so is Not backdated)
  struct bidir_caldata  [pan-server structure]
    - contains only a calibstate for each server
    - unlike perfdata, this structure is created (create_calib)
      only when calibration is activated (typically never) and when
      deactivated is destroyed by radclock_main.c:destroy_calib() .

asym_calibration()
Top level calibration function, called within process_stamp
to process a single new stamp, provided calibration is activated.
It uses BL_track to detect jumps in either OWD, and compares them
against the most recent previous jump to produce “ICZ” (Inferred
Clear Zones). The best (longest) such zone, and the asymmetry
calculated over it, is recording in a struct bidir_calibstate variable.
ICZ are carefully curated and ‘trimmed’ to stay clear of jump position
errors that would pollute them.
The algorithm uses sliding minima (history_min_slide_value_dbl) when
needed so is O(1) per stamp, with a very small constant.

When calibration is terminated, if successful the best result is
available, and could be used to reset the origin of RADclock`s'
asymmetry tracking sub-algo.
The function self-initializes, and uses BL_track_init() to
initialize its BLtrack_state members, including histories.
All of this is done for each server independently in parallel.

BL_track()
Level shift detection algorithm, called to process a single new stamp.
It combines two kinds of detectors:
 - local, a new approach whose core is implemented by LocalDjumpDetect()
 - sliding min-window based  (as in RADclock’s RTT Upshift tracking)
but applies each in both forward and backward ‘order’, resulting in four
detectors:  D, WD, U, WU.
The result is a detector of change in either direction (Up/Down) with
good FP and FN performance, capable of detecting short duration holes
like fast packets.
See comment block for more details.

LocalDjumpDetect
A state-free algorithm, designed to look in a given window for a
‘raw drop’ signature which is highly specific to a Downshift (D) and
which detects with low delay (< 20 stamps typically).
When applied in reverse order (argument order=0) this works to
detect an Upshift viewed in forward order.
The function requires a reference level to be input to define
what constitutes a drop. The calculation of this is not state-free
and is handled by the calling function (BL_track).
See comment block for more details.

Each of the above outputs verbosity that can be mined using key strings
’Calibrat’, ’Viable’, ’detect’.

The detection/calibration fns above (asym_calibration,BL_track,
LocalDjumpDetect) are an implementation of a design developed initially
in matlab (see commits in clockanalysis repo for more detail).

Selective Fake SA suppression
ntp_client()  :
  Installed a SHM suppression hack for selected OCNs (currently
  OCN_SYD and OCN_SYD_2), to blank the transmitted inband trust
  word which is currently carrying noisy fake SA warnings.
  This is important to ensure stamps are accepted for OCN calibration,
  for good behaviour of asymmetry tracking post–calibration, and for
  any OCN serving customers.

SHMalgo :
  Commented out fake SA generation for OCN_SYD to enable clearer
  visibility into TN<-->OCN_SYD  paths conditions as a prelude to
  OCN_SYD_* calibration, and Ext based asymmetry measurement on that
  path.
  For similar reasons, a related hack blocks warnings conerning ICNs
  from influencing the TN’s servertrust: these TN<—>ICN paths have
  reference timing at each end which should be exploited, not obscured.

Such hacks should be dropped when SHM is no longer fake, although warning
suppression during calibration is probably still advisable.

Various ws compliance.