FF/RAD:  Bug fix for FAST flag processing
sysclock_snap2bintime:
The value of the fast option under case SYSCLOCK_FB was not negated,
so fast activation was always the opposite of that intended.
[spotted when identical timestamps were observed for events spaced
 under 1ms apart]
Fixed under both CURRENT and FreeBSD-14.1 .

test_clockcompare:
Tweaks to improve output and to better display/test for the impact
of the FAST flag for both FF and FB families.