NTC:  Support for asymmetry calibration and recording
Establishment of system to allow per-server path asymmetry
to be recorded in the conf file, read and used by RADclock,
and set by calibration.  Involves the retiring of the existing
pan-server system using asym_net, and replacing the
existing time_server parameter with a structure (config_mgr.h) :

struct time_server_Config {
	char domain[MAXLINE];
	double asym;
};

that contains the original time_server domain string and
an associated asymmetry.   The system largely follows the
structure of the NTC_ID database management system,
and the latter was also neatened in the process with improved
comments (was obscure) and removing of longstanding orphan
verbosity on startup.
The system is backward compatible wrt conf files with no asymmetry
value following the server IP/name.  When re-writing the file,
a default of zero is added for each listed server.

An additional CONFIG_TIME_SERVER entry was added
to light reload, restricted to allowing an update to be read
for asymmetry only, for an EXiSting server.
If a change is detected, the new dedicated function
update_config_asym  does the hard work, and provides
old —> new  asym verbosity.

The main changes are in config_mgr.{h,c}, with induced
updates in client_ntp.c, stampinput-livepcap.c, and
process_stamp.

Light Reload
code in signal_handler made more systematic in to make it clearer
what initial processing is required for light options, and to
make verbosity minable by ‘light’.

In config_parse(), the rewriting of the current file in the case of a
PACKAGE_VERSION  mismatch is now suppressed (commented out).
It causes confusing churn in the conf file and is not important.

Connection of the new asym parameters to RADclock:

This is made in process_stamp and sync_bidir.c
The interpretation is that there is a base knowledge of asymmetry,
held by new state variable  state->base_asym , that is always
incorporated into the absolute clock (is added to th_naive estimates,
and which corresponds to the values read and set in the conf file.
Thus the existing state variable Asymhat is
reinterpreted as a ‘delta’ from the base value, that will be used
to effect integration with asymmetry tracking, and/or
asymmetry management (yet to be implemented).
Current changes are simply to update base_asym  when calibration
ends (and update to conf file via update_config_asym), and
to record the corresponding time index, and incorporation
of base_asym into th_naive.

Connection to telemetry:
A simplistic connection is effected by redefining the asymmetry
variable “uA” of a telemetry bundle to be  Asymhat + base_asym.
The benefit is that any change in base_asym (eg at conclusion
of calibration) can be visualized [though it will not appear at
the correct time typically].

To allow easier tracking of per-server behaviour, a sID member
is added to alsostate, and for consistency, to perfstate, matching
the sID member in calibstate.

Repositioning of destroy calibdata code to ensure last asym values
are saved if calibration still active when RADclock is shutdown.

Bug fixes
- fake SA suppression (ntp_client). Second part wasn’t guarded by
is_tn, now both parts are explicitly